feat(theme-builder): move utils.ts into core/
