+++ packages/theme-builder/src/tailwind-shades.ts
@@ -1,6 +1,6 @@
 // imports
 //
-import { Prettify } from './utils';
+import { type Prettify } from './core/utils';
 
 import {
   type Color,
